refactor(form-control): apply createContext util
