+++ packages/bezier-react/src/components/Forms/FormGroup/FormGroup.tsx
@@ -3,7 +3,7 @@ import React, { forwardRef } from 'react'
 import useMergeRefs from '~/src/hooks/useMergeRefs'
 import { noop } from '~/src/utils/functionUtils'
 
-import useFormControlContext from '~/src/components/Forms/useFormControlContext'
+import { useFormControlContext } from '~/src/components/Forms/FormControl'
 
 import type FormGroupProps from './FormGroup.types'
