+++ packages/bezier-react/src/components/Forms/index.ts
@@ -1,6 +1,5 @@
 import type { FormComponentProps } from './Form.types'
 import FormFieldSize from './FormFieldSize'
-import useFormControlContext from './useFormControlContext'
 import useFormFieldProps from './useFormFieldProps'
 import useKeyboardActionLockerWhileComposing from './useKeyboardActionLockerWhileComposing'
 
@@ -8,7 +7,6 @@ export type { FormComponentProps }
 export * from './Inputs/constants/CommonImeControlKeys'
 export {
   FormFieldSize,
-  useFormControlContext,
   useFormFieldProps,
   useKeyboardActionLockerWhileComposing,
 }
