--- packages/bezier-react/src/components/Forms/useFormControlContext.ts
@@ -1,9 +1,0 @@
-import { useContext } from 'react'
-
-import { FormControlContext } from '~/src/components/Forms/FormControl'
-
-function useFormControlContext() {
-  return useContext(FormControlContext)
-}
-
-export default useFormControlContext
